api/requirements.txt
fastapi
fastapi-utilities
slowapi
uvicorn
cachetools
google-genai